refactor: add alias resolve test folder
