+++ src/test/kotlin/me/serce/solidity/lang/core/resolve/SolContractResolveTest.kt
@@ -81,69 +81,6 @@ class SolContractResolveTest : SolResolveTestBase() {
         }
   """)
 
-  fun testResolveSymbolAliases() = testResolveBetweenFiles(
-    InlineFile(
-      code = """
-          contract a {}
-                 //x
-      """,
-      name = "a.sol"
-    ),
-    InlineFile("""
-          import {a as A} from "./a.sol";
-                //^
-          contract b is A {}
-                      
-    """)
-  )
-
-  fun testResolveSymbolAliases2() = testResolveBetweenFiles(
-    InlineFile(
-      code = """
-          contract a {}
-                 //x
-      """,
-      name = "a.sol"
-    ),
-    InlineFile("""
-          import {a as A} from "./a.sol";
-          
-          contract b is A {}
-                      //^
-                      
-    """)
-  )
-
-  fun testResolveSymbolAliasesChain() {
-    InlineFile(
-      code = """
-            contract d {}
-        """,
-      name = "d.sol"
-    )
-    InlineFile(
-      code = """
-            import {d as a} from "./d.sol";
-            contract b {}
-        """,
-      name = "b.sol"
-    )
-    testResolveBetweenFiles(
-      InlineFile(
-        code = """
-            import {b as B} from "./b.sol";
-            contract a {}
-                   //x
-        """,
-        name = "a.sol"
-      ),
-      InlineFile("""
-            import {a as A} from "./a.sol";
-                  //^
-      """)
-    )
-  }
-
   fun testResolveUsingImport() = testResolveBetweenFiles(
     InlineFile(
       code = """
