+++ src/test/kotlin/me/serce/solidity/lang/core/resolve/SolEnumResolveTest.kt
@@ -153,129 +153,5 @@ class SolEnumResolveTest : SolResolveTestBase() {
     )
   )
 
-  fun testResolveEnumFromAlias() = testResolveBetweenFiles(
-    InlineFile(
-      code = """
-         pragma solidity ^0.8.26;
-         
-          enum B { A1, A2 }
-             //x
-    """,
-      name = "a.sol"
-    ),
-    InlineFile("""
-        pragma solidity ^0.8.26;
-        
-        import {B as enumB} from "./a.sol";
 
-        contract C {
-            function f() public {
-                enumB.A2;
-                //^
-            }
-       }
-  """)
-  )
-
-  fun testResolveEnumFromAlias2() = testResolveBetweenFiles(
-    InlineFile(
-      code = """
-         pragma solidity ^0.8.26;
-         
-          enum B { A1, A2 }
-                      //x
-    """,
-      name = "a.sol"
-    ),
-    InlineFile("""
-        pragma solidity ^0.8.26;
-        
-        import {B as enumB} from "./a.sol";
-
-        contract C {
-            function f() public {
-                enumB.A2;
-                    //^
-            }
-       }
-  """)
-  )
-
-  fun testResolveEnumFromAliasInInterface2() = testResolveBetweenFiles(
-    InlineFile(
-      code = """
-         pragma solidity ^0.8.26;
-         
-         interface InterfaceI {
-                     //x
-            enum B { A1, A2 }
-         }
-    """,
-      name = "a.sol"
-    ),
-    InlineFile("""
-        pragma solidity ^0.8.26;
-        
-        import {InterfaceI as Types} from "./a.sol";
-
-        contract C {
-            function f() public {
-                Types.B.A2;
-                //^
-            }
-       }
-  """)
-  )
-
-  fun testResolveEnumFromAliasInInterface3() = testResolveBetweenFiles(
-    InlineFile(
-      code = """
-         pragma solidity ^0.8.26;
-         
-         interface InterfaceI {
-            enum B { A1, A2 }
-                       //x
-         }
-    """,
-      name = "a.sol"
-    ),
-    InlineFile("""
-        pragma solidity ^0.8.26;
-        
-        import {InterfaceI as Types} from "./a.sol";
-
-        contract C {
-            function f() public {
-                Types.B.A2;
-                       //^
-            }
-       }
-  """)
-  )
-
-  fun testResolveEnumFromAliasInInterface4() = testResolveBetweenFiles(
-    InlineFile(
-      code = """
-         pragma solidity ^0.8.26;
-         
-         interface InterfaceI {
-            enum B { A1, A2 }
-               //x
-         }
-    """,
-      name = "a.sol"
-    ),
-    InlineFile("""
-        pragma solidity ^0.8.26;
-        
-        import {InterfaceI as Types} from "./a.sol";
-
-        contract C {
-            function f() public {
-                Types.B.A2;
-                    //^
-            }
-       }
-  """)
-  )
 }
