+++ src/test/kotlin/me/serce/solidity/lang/core/resolve/SolEventResolveTest.kt
@@ -182,57 +182,6 @@ class SolEventResolveTest : SolResolveTestBase() {
     )
   )
 
-  fun testResolveEventFromAlias() = testResolveBetweenFiles(
-    InlineFile(
-      code = """
-         pragma solidity ^0.8.26;
-         
-          event Closed();
-                 //x    
-    """,
-      name = "a.sol"
-    ),
-    InlineFile("""
-        pragma solidity ^0.8.26;
-        
-        import {Closed as EventClosed} from "./a.sol";
-
-        contract C {
-            function f() public {
-               emit EventClosed();
-                        //^
-            }
-       }
-  """)
-  )
-
-  fun testResolveEventFromAliasInInterface() = testResolveBetweenFiles(
-    InlineFile(
-      code = """
-         pragma solidity ^0.8.26;
-         
-          interface InterfaceI {
-            event Closed();
-                 //x
-         }   
-    """,
-      name = "a.sol"
-    ),
-    InlineFile("""
-        pragma solidity ^0.8.26;
-        
-        import {InterfaceI as Types} from "./a.sol";
-
-        contract C {
-            function f() public {
-               emit Types.Closed();
-                           //^
-            }
-       }
-  """)
-  )
-
-
   override fun checkByCode(code: String) {
     checkByCodeInternal<SolFunctionCallExpression, SolNamedElement>(code)
   }
