+++ src/test/kotlin/me/serce/solidity/lang/core/resolve/SolStructResolveTest.kt
@@ -129,89 +129,7 @@ class SolStructResolveTest : SolResolveTestBase() {
     )
   )
 
-  fun testResolveStructFromAlias() = testResolveBetweenFiles(
-    InlineFile(
-      code = """
-         pragma solidity ^0.8.26;
-         
-          struct Prop {
-                //x
-              uint prop1;
-              uint prop2;
-          }
-    """,
-      name = "a.sol"
-    ),
-    InlineFile("""
-        pragma solidity ^0.8.26;
-        
-        import {Prop as Proposal} from "./a.sol";
 
-        contract C {
-            Proposal prop = Proposal(0, 1);
-            //^
-            function f() public {
-                prop.prop1;
-            }
-       }
-  """)
-  )
-
-  fun testResolveStructFromAlias2() = testResolveBetweenFiles(
-    InlineFile(
-      code = """
-         pragma solidity ^0.8.26;
-         
-          struct Prop {
-                //x
-              uint prop1;
-              uint prop2;
-          }
-    """,
-      name = "a.sol"
-    ),
-    InlineFile("""
-        pragma solidity ^0.8.26;
-        
-        import {Prop as Proposal} from "./a.sol";
-
-        contract C {
-            Proposal prop = Proposal(0, 1);
-                            //^
-            function f() public {
-                prop.prop1;
-            }
-       }
-  """)
-  )
-
-  fun testResolveStructFromAlias4() = testResolveBetweenFiles(
-    InlineFile(
-      code = """
-         pragma solidity ^0.8.26;
-         
-          struct Prop {
-              uint prop1;
-                  //x
-              uint prop2;
-          }
-    """,
-      name = "a.sol"
-    ),
-    InlineFile("""
-        pragma solidity ^0.8.26;
-        
-        import {Prop as Proposal} from "./a.sol";
-
-        contract C {
-            Proposal prop = Proposal(0, 1);
-            function f() public {
-                prop.prop1;
-                      //^
-            }
-       }
-  """)
-  )
 
   fun testResolveStructFromInterface() = testResolveBetweenFiles(
     InlineFile(
