+++ src/test/kotlin/me/serce/solidity/lang/core/resolve/SolUserDefinedValueTypeResolveTest.kt
@@ -69,54 +69,4 @@ class SolUserDefinedValueTypeResolveTest : SolResolveTestBase() {
     """
     )
   )
-
-  fun testUserDefinedValueTypeResolveAlias() = testResolveBetweenFiles(
-    InlineFile(
-      code = """
-         pragma solidity ^0.8.26;
-         
-         type Decimal18 is uint256;
-                //x
-    """,
-      name = "a.sol"
-    ),
-    InlineFile(
-      """
-        pragma solidity ^0.8.26;
-        
-        import {Decimal18 as customType} from "./a.sol";
-
-        contract C {
-            function transfer(address to, customType value) external;
-                                            //^
-       }
-  """
-    )
-  )
-
-  fun testUserDefinedValueTypeResolveAliasInInterface() = testResolveBetweenFiles(
-    InlineFile(
-      code = """
-         pragma solidity ^0.8.26;
-         
-         interface InterfaceI {
-           type Decimal18 is uint256;
-                  //x
-         }
-    """,
-      name = "a.sol"
-    ),
-    InlineFile(
-      """
-        pragma solidity ^0.8.26;
-        
-        import {InterfaceI as Types} from "./a.sol";
-
-        contract C {
-            function transfer(address to, Types.Decimal18 value) external;
-                                                //^
-       }
-  """
-    )
-  )
 }
