+++ src/test/kotlin/me/serce/solidity/lang/core/resolve/alias/SolAliasInterfaceResolveTest.kt
@@ -1,0 +1,31 @@
+package me.serce.solidity.lang.core.resolve.alias
+
+import me.serce.solidity.lang.core.resolve.SolResolveTestBase
+
+class SolAliasInterfaceResolveTest : SolResolveTestBase() {
+  fun testResolveInterfaceFromAlias() = testResolveBetweenFiles(
+    InlineFile(
+      code = """
+         pragma solidity ^0.8.26;
+         
+         interface InterfaceI {
+                     //x
+            enum B { A1, A2 }
+         }
+    """,
+      name = "a.sol"
+    ),
+    InlineFile("""
+        pragma solidity ^0.8.26;
+        
+        import {InterfaceI as Types} from "./a.sol";
+
+        contract C {
+            function f() public {
+                Types.B.A2;
+                //^
+            }
+       }
+  """)
+  )
+}
